Add Test Support Lib; Initial Work on REPL
Moving test logging out to a new module testsupport, which acts as a primitive unit testing environment. In tandem with this, the Makefile has been updated with a watcher task to support TDD.
I've alse added an 'appsupport' module which will contain things like command parsing, which should be tested independently of a repl.